Moved helper functions to new file
